+++ ZooUiProject/src/app/features/animal/ration/ration.component.ts
@@ -2,12 +2,14 @@ import { Component, OnInit, OnDestroy } from '@angular/core';
 import { IRation } from 'src/app/core/interfaces/ration.interface';
 import { FormGroup, FormBuilder, Validators } from '@angular/forms';
 import { Subject } from 'rxjs';
-import { ActivatedRoute } from '@angular/router';
-import { configureToastr, getButtonStateImport, hasPatternErrorImport, hasCustomErrorImport } from 'src/app/core/helpers';
+import { ActivatedRoute, Router } from '@angular/router';
+import { configureToastr, getButtonStateImport, hasPatternErrorImport, hasCustomErrorImport, deleteConfirmImport } from 'src/app/core/helpers';
 import { ToastrService } from 'ngx-toastr';
 import { RationService } from '../common/ration.service';
 import { takeUntil, finalize } from 'rxjs/operators';
 import { enums } from 'src/app/core/constants/enums';
+import { AnimalService } from '../common/animal.service';
+import { IAnimal } from 'src/app/core/interfaces/animal.interface';
 
 @Component({
   selector: 'app-ration',
@@ -16,22 +18,26 @@ import { enums } from 'src/app/core/constants/enums';
 })
 export class RationComponent implements OnInit, OnDestroy {
 
+  animal: IAnimal;
   rations: IRation[];
   rationForm: FormGroup;
+  rationId: number;
 
   private update = false;
   private animalId: number;
-  private rationId: number;
   private destroy$ = new Subject<any>();
 
   constructor(private fb: FormBuilder,
               private route: ActivatedRoute,
+              private router: Router,
               private toastr: ToastrService,
-              private rationService: RationService) { }
+              private rationService: RationService,
+              private animalSerivce: AnimalService) { }
 
   ngOnInit(): void {
-    this.animalId = +this.route.snapshot.params.id;
+    this.animalId = +this.route.snapshot.params.animalId;
     this.createForm();
+    this.getAnimal();
     this.getRations();
     configureToastr(this.toastr);
   }
@@ -39,8 +45,14 @@ export class RationComponent implements OnInit, OnDestroy {
   createForm(): void {
     this.rationForm = this.fb.group({
       foodName: [null, [Validators.required, Validators.maxLength(100)]],
-      description: [null, [Validators.maxLength(200)]]
+      description: [null, [Validators.required, Validators.maxLength(200)]]
     });
+  }
+
+  getAnimal() {
+    this.animalSerivce.getAnimal(this.animalId)
+      .pipe(takeUntil(this.destroy$))
+      .subscribe(data => this.animal = data);
   }
 
   getRations() {
@@ -54,13 +66,15 @@ export class RationComponent implements OnInit, OnDestroy {
 
     this.update = true;
     this.rationId = ration.rationId;
-    this.rationForm.get('foodName').setValue(ration.foodName);
-    this.rationForm.get('description').setValue(ration.description);
+    this.rationForm.patchValue({
+      foodName: ration.foodName,
+      description: ration.description
+    });
   }
 
   onSubmit(): void {
     if (this.rationForm.valid) {
-      this.update ? this.createRation() : this.updateRation();
+      this.update ? this.updateRation() : this.createRation();
     } else {
       this.rationForm.markAllAsTouched();
     }
@@ -109,6 +123,10 @@ export class RationComponent implements OnInit, OnDestroy {
   }
 
   deleteRation(id) {
+    if (!deleteConfirmImport(this.rations.find(x => x.rationId === id).foodName)) {
+      return;
+    }
+
     this.rationService.deleteRation(id)
       .pipe(
         finalize(() => this.getRations()),
@@ -120,6 +138,10 @@ export class RationComponent implements OnInit, OnDestroy {
           console.log(err);
         }
       );
+  }
+
+  backToAnimals() {
+    this.router.navigate(['/animal/details', this.animalId]);
   }
 
   resetForm(): void {
